- namespace normalization refactoring   for:      "entityframeworkrepository.cache";      "entityframeworkrepository.repository";      "entityframeworkrepository.repository.cache";
