+++ entityframeworkrepository.cache/AbstractCacheProvider.cs
@@ -16,6 +16,5 @@ namespace entityframeworkrepository.cache
         public abstract bool Set(ICacheKey cacheKey, object value, ICachePolicy cachePolicy);
         public abstract object GetOrAdd(ICacheKey cacheKey, Func<ICacheKey, object> valueFactory, ICachePolicy cachePolicy);
         public abstract Task<object> GetOrAddAsync(ICacheKey cacheKey, Func<ICacheKey, Task<object>> valueFactory, ICachePolicy cachePolicy);
-      
     }
 }
