+++ entityframeworkrepository.cache/MemoryCacheProvider.cs
@@ -5,6 +5,7 @@ using System.Runtime.Caching;
 using System.Runtime.Caching.Generic;
 using System.Threading.Tasks;
 
+
 namespace entityframeworkrepository.cache
 {
     /// <summary>
@@ -89,7 +90,7 @@ namespace entityframeworkrepository.cache
         /// The value for the key. This will be either the existing value for the key if the key is already in the cache,
         /// or the new value for the key as returned by <paramref name="valueFactory"/> if the key was not in the cache.
         /// </returns>
-        public async override Task<object> GetOrAddAsync(ICacheKey cacheKey, Func<ICacheKey, Task<object>> valueFactory, ICachePolicy cachePolicy)
+        public override async Task<object> GetOrAddAsync(ICacheKey cacheKey, Func<ICacheKey, Task<object>> valueFactory, ICachePolicy cachePolicy)
         {
             var key = GetKey(cacheKey);
             var cachedResult = MemoryCache.Default.Get(key);
