+++ entityframeworkrepository.cache/Query/LocalCollectionExpander.cs
@@ -9,7 +9,7 @@ namespace entityframeworkrepository.cache.Query
     /// <summary>
     /// Enables cache key support for local collection values.
     /// </summary>
-    internal class LocalCollectionExpander : ExpressionVisitor
+    internal partial class LocalCollectionExpander : ExpressionVisitor
     {
         public static Expression Rewrite(Expression expression)
         {
@@ -57,26 +57,6 @@ namespace entityframeworkrepository.cache.Query
             var printer = Activator.CreateInstance(printerType, value);
 
             return Expression.Constant(printer);
-        }
-
-        /// <summary>
-        /// Overrides ToString to print each element of a collection.
-        /// </summary>
-        /// <remarks>
-        /// Inherits List in order to support List.Contains instance method as well
-        /// as standard Enumerable.Contains/Any extension methods.
-        /// </remarks>
-        class Printer<T> : List<T>
-        {
-            public Printer(IEnumerable collection)
-            {
-                this.AddRange(collection.Cast<T>());
-            }
-
-            public override string ToString()
-            {
-                return "{" + this.ToConcatenatedString(t => t.ToString(), "|") + "}";
-            }
         }
     }
 }
