+++ entityframeworkrepository.core/cache/CacheExpirationMode.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace entityframeworkrepository.cache
 {
     /// <summary>
