+++ entityframeworkrepository.core/cache/ICacheTag.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-// ReSharper disable CheckNamespace
+
 
 namespace entityframeworkrepository.cache
 {
